# Hand-roll faster and safer version of `import/no-restricted-paths`