--- a/eslint-local-rules/noCrossBoundaryImports.js
+++ b/eslint-local-rules/noCrossBoundaryImports.js
@@ -18,7 +18,6 @@
 const path = require("node:path");
 const multimatch = require("multimatch");
 
-// eslint-disable-next-line unicorn/prefer-module
 module.exports = {
   meta: {
     type: "problem",