--- a/knip.mjs
+++ b/knip.mjs
@@ -74,6 +74,8 @@ const knipConfig = {
     "@types/gapi.client.oauth2-v2",
     // Used by Code Editor so format on save matches pre-commit behavior
     "prettier",
+    // Used by eslint-local-rules/noCrossBoundaryImports.js
+    "minimatch",
   ],
   rules: {
     // Issue Type reference: https://knip.dev/reference/issue-types/