--- a/package.json
+++ b/package.json
@@ -268,6 +268,7 @@
     "mockdate": "^3.0.5",
     "msw": "^1.3.2",
     "msw-storybook-addon": "^1.10.0",
+    "multimatch": "^5.0.0",
     "node-polyfill-webpack-plugin": "^3.0.0",
     "prettier": "3.1.0",
     "react-select-event": "^5.3.0",