## What does this PR do?

- Closes https://github.com/pixiebrix/eslint-config-pixiebrix/issues/35
- Closes https://github.com/pixiebrix/eslint-config-pixiebrix/issues/123

👆 😅 it was so slow that I opened two issues for the same thing over the years. 

## Before (2m 20s)

```
Rule                                            | Time (ms) | Relative
:-----------------------------------------------|----------:|--------:
import/no-restricted-paths                      | 50222.336 |    49.1%
@typescript-eslint/no-unsafe-assignment         |  8793.188 |     8.6%
unicorn/no-unnecessary-polyfills                |  4247.493 |     4.1%
@typescript-eslint/no-floating-promises         |  3331.880 |     3.3%
@typescript-eslint/no-confusing-void-expression |  2293.998 |     2.2%
@typescript-eslint/no-unsafe-argument           |  1524.524 |     1.5%
import/no-self-import                           |  1359.165 |     1.3%
@typescript-eslint/no-redeclare                 |  1191.053 |     1.2%
react/no-direct-mutation-state                  |   915.010 |     0.9%
unicorn/expiring-todo-comments                  |   900.622 |     0.9%
```

## After (1m 30s) 

```
Rule                                            | Time (ms) | Relative
:-----------------------------------------------|----------:|--------:
@typescript-eslint/no-unsafe-assignment         |  9179.275 |    17.2%
import/no-self-import                           |  4934.737 |     9.2%
unicorn/no-unnecessary-polyfills                |  4443.853 |     8.3%
@typescript-eslint/no-floating-promises         |  3233.091 |     6.1%
@typescript-eslint/no-confusing-void-expression |  2376.571 |     4.5%
@typescript-eslint/no-unsafe-argument           |  1567.075 |     2.9%
@typescript-eslint/no-redeclare                 |  1303.585 |     2.4%
react/no-direct-mutation-state                  |   941.887 |     1.8%
unicorn/prevent-abbreviations                   |   934.537 |     1.8%
unicorn/expiring-todo-comments                  |   920.177 |     1.7%
```

## Checklist

- [x] Add tests: 🚀 
- [x] Designate a primary reviewer: @grahamlangford