@ljharb 2024-02-10
@fregante fwiw if this actually is a faster 1:1 replacement for the import rule, i'd love an upstream PR :-)