@fregante 2024-02-11
Unfortunately this rule takes a lot of shortcuts and just compares the top-level folder. It also changes how the comparison is done (one rule per folder vs "list of folders that can't import from each other")